# ISPN-6580 Hotrod performance regressions after ISPN-5342 ISPN-6545