--- a/server/hotrod/src/main/java/org/infinispan/server/hotrod/CacheDecodeContextLog.java
+++ b/server/hotrod/src/main/java/org/infinispan/server/hotrod/CacheDecodeContextLog.java
@@ -1,0 +1,12 @@
+package org.infinispan.server.hotrod;
+
+import org.infinispan.server.core.logging.JavaLog;
+import org.infinispan.util.logging.LogFactory;
+
+/**
+ * @author wburns
+ * @since 9.0
+ */
+public class CacheDecodeContextLog {
+   static final JavaLog log = LogFactory.getLog(CacheDecodeContext.class, JavaLog.class);
+}