… client/hotrod-client/src/test/java/org/infinispan/client/hotrod/stress/RemoteClientPutGetTest.java
… client/hotrod-client/src/test/java/org/infinispan/client/hotrod/test/HotRodClientTestingUtil.java
… server/core/src/main/scala/org/infinispan/server/core/UnsignedNumeric.java
… server/hotrod/src/main/java/org/infinispan/server/hotrod/ResponseWriting.java
… server/hotrod/src/main/java/org/infinispan/server/hotrod/logging/HotRodAccessLoggingHandler.java
… server/hotrod/src/main/scala/org/infinispan/server/hotrod/AbstractEncoder1x.scala
… server/hotrod/src/main/scala/org/infinispan/server/hotrod/AbstractVersionedDecoder.scala
… server/hotrod/src/main/scala/org/infinispan/server/hotrod/Decoder10.scala
… server/hotrod/src/main/scala/org/infinispan/server/hotrod/Decoder2x.scala
… server/hotrod/src/main/scala/org/infinispan/server/hotrod/Encoder2x.scala
… server/hotrod/src/main/scala/org/infinispan/server/hotrod/HotRodEncoder.scala
… server/hotrod/src/main/scala/org/infinispan/server/hotrod/OperationStatus.scala
… server/hotrod/src/main/scala/org/infinispan/server/hotrod/Response.scala
… server/hotrod/src/main/scala/org/infinispan/server/hotrod/transport/HotRodChannelInitializer.scala