https://issues.jboss.org/browse/ISPN-6580

Some numbers from the ```RemoteClientPutGetTest```, on Linux (server run with ```-Dinfinispan.server.channel.epoll=true```) 

||8.2.1.Final | 9.0.0-SNAPSHOT
----|-------|-----------------------
|Put| 13s | 15s 
|Get| 71s | 66s
|Alloc| 130 MB/s | 47 MB/s