@gustavonalle 2016-05-11
Updated again. Pushed one tiny fix to ```RemoteClientPutGetTest```