@gustavonalle 2016-05-11
Rebased. Also a slight change of plan: epoll is disabled by default for now, and can be enabled by a System property.